[Refactor] useFetchHpMsg hook 수정
